+++ src/main/java/ddip/me/ddipbe/presentation/dto/MemberController.java
@@ -5,6 +5,10 @@ import ddip.me.ddipbe.domain.Member;
 import ddip.me.ddipbe.global.annotation.SessionMemberId;
 import ddip.me.ddipbe.global.dto.ResponseEnvelope;
 import ddip.me.ddipbe.global.util.SessionUtil;
+import ddip.me.ddipbe.presentation.dto.request.SigninRequest;
+import ddip.me.ddipbe.presentation.dto.request.SignupRequest;
+import ddip.me.ddipbe.presentation.dto.response.MemberIdResponse;
+import ddip.me.ddipbe.presentation.dto.response.MemberMeResponse;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
