+++ src/main/java/ddip/me/ddipbe/presentation/dto/request/SigninRequest.java
@@ -1,9 +1,9 @@
-package ddip.me.ddipbe.presentation.dto;
+package ddip.me.ddipbe.presentation.dto.request;
 
 import lombok.Data;
 
 @Data
-public class SignupRequest {
+public class SigninRequest {
 
     private String email;
     private String password;
